+++ src/setupTests.ts
@@ -3,3 +3,7 @@ import { TextEncoder, TextDecoder } from 'util';
 Object.assign(global, { TextEncoder, TextDecoder });
 
 import '@testing-library/jest-dom';
+
+beforeEach(() => {
+  localStorage.clear();
+});